backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
python-multipart
boto3
pypdf
magnum